pyproject.toml
[tool.poetry]
name = "img_data_gen"
version = "0.1.0"
description = ""
authors = ["Sven Becker <Sven.Becker@mgm-tp.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.2.0"
numpy = "^1.23.1"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
